LayoutTests/platform/glib/http/tests/websocket/tests/hybi/long-invalid-header-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/long-invalid-header' failed: Server ignored WebSocket handshake
Make sure WebSocket gives errors on long invalid upgrade header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket is closed
PASS successfullyParsed is true

TEST COMPLETE
